requirements.txt
flask
numpy
pillow
faiss-cpu
tqdm
scipy
werkzeug